requirements.txt
# nucleu data / ML
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0

# NLP & embeddings
nltk>=3.9
spacy>=3.8.0
sentence-transformers>=3.0.0

# vizualizare & dashboards
matplotlib>=3.8.0
seaborn>=0.13.0
wordcloud>=1.9.0
streamlit>=1.55.0

# utilitare & API-uri
python-dotenv>=1.0.0
requests>=2.32.0
tweepy>=4.0.0
praw>=7.0.0

# optional, pentru lucru cu timp si date
pyyaml>=6.0.0
python-dateutil>=2.9.0
